requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
greenlet==3.1.1
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
aiomysql==0.2.0
pymysql==1.1.1
python-dotenv==1.0.1
email-validator==2.2.0




